app/agents/core_agents/llm_agent.py
import logging
from typing import Any, Dict, List, Optional
from pydantic import BaseModel, Field

from app.agents.core_agents.agent_types import Agent
from app.models.agent.belief import Belief
from app.models.agent.desire import Desire
from app.models.agent.intention import Intention
from app.models.agent.goal import Goal
from app.models.message import ACLMessage, Performative
from app.services.llm_service import LLMService
from app.tools.llm_manager import LLMManager
from app.tools.reasoning_engine import ReasoningEngine
from app.tools.reasoner import Reasoner
from app.tools.ontology_reasoner import OntologyReasoner

logger = logging.getLogger(__name__)

class LLMAgent(Agent):
    llm_service: LLMService = Field(default_factory=LLMService)
    llm_manager: LLMManager = Field(default_factory=LLMManager)
    reasoning_engine: ReasoningEngine = Field(default_factory=ReasoningEngine)
    reasoner: Reasoner = Field(default_factory=Reasoner)
    ontology_reasoner: OntologyReasoner = Field(default_factory=OntologyReasoner)

    class Config:
        arbitrary_types_allowed = True

    async def process_message(self, message: ACLMessage):
        """Process incoming messages using LLM capabilities."""
        try:
            if message.performative == Performative.REQUEST:
                response = await self.handle_request(message.content)
            elif message.performative == Performative.INFORM:
                response = await self.process_information(message.content)
            elif message.performative == Performative.QUERY:
                response = await self.handle_query(message.content)
            else:
                response = "Unsupported message performative"

            await self.agent_communication_service.send_message(
                self.agent_id,
                message.sender_id,
                Performative.INFORM,
                response
            )
        except Exception as e:
            logger.error(f"Error processing message: {str(e)}")

    async def handle_request(self, request: str) -> str:
        """Handle requests using LLM capabilities."""
        prompt = f"Given the following request: '{request}', how should I respond or act?"
        return await self.llm_service.generate_agent_response(self, prompt)

    async def process_information(self, information: str) -> str:
        """Process information using LLM capabilities."""
        prompt = f"Given the following information: '{information}', what are the key points or actions to take?"
        return await self.llm_service.generate_agent_response(self, prompt)

    async def handle_query(self, query: str) -> str:
        """Handle queries using LLM capabilities."""
        prompt = f"Given the following query: '{query}', what is the appropriate response?"
        return await self.llm_service.generate_agent_response(self, prompt)

    async def reason(self, context: Dict[str, Any]) -> Dict[str, Any]:
        """Perform reasoning based on the given context."""
        reasoning_result = await super().reason(context)
        
        # Use reasoning engine for advanced decision-making
        reasoning_engine_result = await self.reasoning_engine.reason(context)
        
        # Use reasoner for logical inference
        reasoner_result = self.reasoner.infer(context)
        
        # Use ontology reasoner for knowledge inference
        ontology_result = await self.ontology_reasoner.infer_knowledge(context)
        
        # Combine results from different reasoning methods
        combined_result = {
            **reasoning_result,
            **reasoning_engine_result,
            **reasoner_result,
            **ontology_result
        }
        
        # Generate new beliefs, desires, and intentions based on combined reasoning
        new_beliefs = await self.llm_service.generate_beliefs(combined_result)
        new_desires = await self.llm_service.generate_desires(combined_result)
        new_intentions = await self.llm_service.generate_intentions(combined_result)
        
        self.update_beliefs(new_beliefs)
        self.update_desires(new_desires)
        self.update_intentions(new_intentions)
        
        return combined_result

    async def act(self) -> None:
        """Perform the agent's main action cycle."""
        context = self.get_current_state()
        await self.reason(context)
        for intention in self.intentions:
            if intention.status == "active":
                await self.execute_intention(intention)

    async def execute_intention(self, intention: Intention) -> None:
        """Execute a given intention."""
        # Implementation depends on the specific intentions of the LLM agent
        pass

    def get_current_state(self) -> Dict[str, Any]:
        """Get the current state of the agent."""
        return {
            "beliefs": [belief.dict() for belief in self.beliefs],
            "desires": [desire.dict() for desire in self.desires],
            "intentions": [intention.dict() for intention in self.intentions],
        }